requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.24.0
telepot==12.7
